+++ core/database/src/main/java/com/mifos/room/entities/zipmodels/ClientAndClientAccounts.kt
@@ -9,8 +9,8 @@
  */
 package com.mifos.room.entities.zipmodels
 
-import com.mifos.core.entity.client.Client
 import com.mifos.room.entities.accounts.ClientAccounts
+import com.mifos.room.entities.client.Client
 
 /**
  * Model for Observable.zip. This Model used to combine the Client and ClientAccount in response
