+++ core/domain/src/main/java/com/mifos/core/domain/useCases/GetClientIdentifierTemplateUseCase.kt
@@ -11,7 +11,7 @@ package com.mifos.core.domain.useCases
 
 import com.mifos.core.common.utils.Resource
 import com.mifos.core.data.repository.ClientIdentifierDialogRepository
-import com.mifos.core.objects.noncoreobjects.IdentifierTemplate
+import com.mifos.core.model.objects.noncoreobjects.IdentifierTemplate
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
