+++ core/domain/src/main/java/com/mifos/core/domain/useCases/GetClientIdentifiersUseCase.kt
@@ -11,7 +11,7 @@ package com.mifos.core.domain.useCases
 
 import com.mifos.core.common.utils.Resource
 import com.mifos.core.data.repository.ClientIdentifiersRepository
-import com.mifos.core.objects.noncoreobjects.Identifier
+import com.mifos.core.model.objects.noncoreobjects.Identifier
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
